Move QuerySet to its own module
